API and API editor changes
In the RealityFlowAPIEditor.cs I added editor buttons that display the utility of the API functions.

In the RealityFlowAPI.cs I updated the spawn and deletion functions, such that deletion works for peer spawning.  I also added functions for finding spawnedObjects, updating objects transform component, adding prefabs to the prefab catalogue, and adding game objects to the prefab catalogue.

In the NetworkSpawnManager.cs I added multiple debug lines and made some tweaks to the ResolveIndex function as well as helper functions to interact with the spawnedForPeer and spawnedForRoom dictionaries.