Implementation of Mute feature
As of right now these changes make it so when you click record audio button to speak to Whisper your mic is muted. This could be extended to provide users with a mute button so peers won't be able to hear what they are saying when they are muted. This likely requires further testing as the only testing I was able to conduct was between a built application and the unity editor.