Changes
Updated the LLM window and system. Added exit buttons to the menus. Down on the left stick trigger the opening and closing of the LLM window holding the secondary button on the left controller records the voice and starts the llm process. Pressing the primary button on the left controller executes code.  Updated action logger to handle multiple action execution at one time. Reorganized Action log and RealityFlowAPI script accordingly.

Also some early code for the sending of screenshots to the LLM as context. The main scene for me is now LLMSceneWithWorkingMRTK